+++ analysis/src/Uri.mli
@@ -6,3 +6,4 @@ val stripPath : bool ref
 val toPath : t -> string
 val toString : t -> string
 val toTopLevelLoc : t -> Location.t
+val encodeURIComponent : string -> string