+++ CHANGELOG.md
@@ -12,6 +12,18 @@
 
 ## master
 
+#### :rocket: New Feature
+
+- Add support for prop completion for JSX V4 https://github.com/rescript-lang/rescript-vscode/pull/579
+- Add support for create interface file for JSX V4 https://github.com/rescript-lang/rescript-vscode/pull/580
+- Expand one level of type definition on hover. Dig into record/variant body. https://github.com/rescript-lang/rescript-vscode/pull/584
+- Add clickable links to type definitions in hovers. https://github.com/rescript-lang/rescript-vscode/pull/585
+- Add experimental signature help for function calls. https://github.com/rescript-lang/rescript-vscode/pull/547
+
+#### :bug: Bug Fix
+
+- Fix printing of record types with optional fields https://github.com/rescript-lang/rescript-vscode/pull/584
+
 ## v1.6.0
 
 #### :rocket: New Feature
@@ -23,6 +35,7 @@
 
 #### :bug: Bug Fix
 
+- Fix issue where debug output would end up in the JSON file produced by Reanalyze https://github.com/rescript-lang/rescript-vscode/pull/575
 - Fix issue where autocomplete would not perform type instantiation https://github.com/rescript-lang/rescript-vscode/pull/561
 - Fix issue where hovering over a field in record construction would show the type without instantiating its type arguments https://github.com/rescript-lang/rescript-vscode/pull/560
 - Fix Incorrect semantic highlighting of `external` declarations https://github.com/rescript-lang/rescript-vscode/pull/517